les dernieres modification

+++ smart_morocco/src/main/java/com/example/smart_morocco/repository/PackRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PackRepository extends JpaRepository<Pack, Integer> {
+public interface PackRepository extends JpaRepository<Pack, Long> {
 }
